+++ core/lib/Thelia/Controller/Admin/TranslationsController.php
@@ -181,7 +181,7 @@ class TranslationsController extends BaseAdminController
                 case 'co':
                     $directory = THELIA_LIB;
                     $domain = 'core';
-                    $i18n_directory = THELIA_LIB . 'Config' . DS . 'I18n';
+                    $i18nDirectory = THELIA_LIB . 'Config' . DS . 'I18n';
                     $walkMode = TranslationEvent::WALK_MODE_PHP;
                     break;
 
@@ -247,6 +247,12 @@ class TranslationsController extends BaseAdminController
                                 true
                             );
 
+                            $event
+                                ->setDomain($domain)
+                                ->setLocale($this->getCurrentEditionLocale())
+                                ->setCustomFallbackStrings($this->getRequest()->get('translation_custom', []))
+                                ->setGlobalFallbackStrings($this->getRequest()->get('translation_global', []));
+
                             $this->getDispatcher()->dispatch(TheliaEvents::TRANSLATION_WRITE_FILE, $event);
 
                             if ($save_mode == 'stay') {
@@ -272,7 +278,7 @@ class TranslationsController extends BaseAdminController
                 $this->getDispatcher()->dispatch(TheliaEvents::TRANSLATION_GET_STRINGS, $event);
 
                 // Estimate number of fields, and compare to php ini max_input_vars
-                $stringsCount = $event->getTranslatableStringCount() * 2 + 6;
+                $stringsCount = $event->getTranslatableStringCount() * 4 + 6;
 
                 if ($stringsCount > ini_get('max_input_vars')) {
                     $templateArguments['max_input_vars_warning']  = true;
@@ -281,10 +287,29 @@ class TranslationsController extends BaseAdminController
                 } else {
                     $templateArguments['all_strings'] = $event->getTranslatableStrings();
                 }
+
+                $templateArguments['is_writable'] = $this->checkWritableI18nDirectory(THELIA_LOCAL_DIR . 'I18n');
             }
         }
 
         return $this->render('translations', $templateArguments);
+    }
+
+    /**
+     * Check if a directory is writable or if the parent directory is writable
+     *
+     * @param string $dir   the directory to test
+     * @return boolean      return true if the directory is writable otr if the parent dir is writable.
+     */
+    public function checkWritableI18nDirectory($dir)
+    {
+        if (file_exists($dir)) {
+            return is_writable($dir);
+        }
+
+        $parentDir = dirname($dir);
+
+        return file_exists($parentDir) && is_writable($parentDir);
     }
 
     public function defaultAction()